refactor(smg): move mesh name validation into config::validation
Bin entrypoints stay test-free: the check moves to a pub
validate_mesh_server_name helper tested beside the other config
validators; main.rs keeps a one-line call.
